requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
requests
openai
google-api-python-client


